pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jinguzhou"
version = "0.1.0"
description = "A model-agnostic safety gateway, policy engine, and audit layer for LLM applications."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Jinguzhou Contributors" }
]
keywords = ["llm", "ai-safety", "gateway", "policy-engine", "audit"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "pydantic>=2.7,<3.0",
  "pyyaml>=6.0,<7.0",
  "typer>=0.12,<1.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
jinguzhou = "jinguzhou.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
